src/requirements.txt
django~=3.2
uWSGI~=2.0
psycopg2~=2.9
PyYAML~=5.3.1
djangorestframework~=3.14
Markdown~=3.3
django-filter~=21.1
django-rest-framework-social-oauth2==1.1.0
dj-database-url~=2.1.0
arrow~=1.2.3
uritemplate~=4.1
